+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Investigate option in context menu to send indicators to the popup
 - Icon to mark the tools that support the automatic graph population feature
 - Icon to mark the tools that require user interaction
+- Icon to mark the tools that require an account
 
 ### Change
 - Updated browser-polyfill library